+++ packages/hydrogen/src/entry-server.tsx
@@ -257,7 +257,7 @@ function buildReactApp({
       context={context}
     >
       <HelmetProvider context={helmetContext}>
-        <App request={request} response={componentResponse} {...props} />
+        <App {...props} request={request} response={componentResponse} />
       </HelmetProvider>
     </StaticRouter>
   );